cours_en_ligne/requirements.txt
absl-py==2.1.0
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bleach==6.0.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.2
cycler==0.11.0
Cython==3.0.10
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.2.0
Django==3.2.25
django-bootstrap4==23.1
django-braces==1.15.0
django-ckeditor==6.3.2
django-crispy-forms==2.0
django-embed-video==1.4.10
django-heroku==0.3.1
django-js-asset==2.0.0
django-memcache-status==2.2
django-rest-framework==0.1.0
django-tinymce==3.6.1
django-widget-tweaks==1.4.12
djangorestframework==3.15.1
entrypoints==0.4
exceptiongroup==1.2.1
fastjsonschema==2.19.1
feedparser==6.0.11
flatbuffers==24.3.25
fonttools==4.38.0
frozenlist==1.3.3
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.8.0
idna==3.7
importlib-metadata==2.1.3
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
jaraco.context==4.3.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.17.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
kaggle==1.6.12
keras==2.11.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==9.1.0
multidict==6.0.5
nbclassic==1.1.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.6.0
nltk==3.8.1
notebook==6.5.7
notebook_shim==0.2.4
npm==0.1.1
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optional-django==0.1.0
packaging==24.0
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.17.1
prompt_toolkit==3.0.45
protobuf==3.19.6
psutil==5.9.8
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyjokes==0.6.0
PyJWT==2.8.0
pyparsing==3.1.2
pypiwin32==223
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-memcached==1.62
python-slugify==8.0.4
pyttsx3==2.90
pytz==2024.1
pywin32==306
pywinpty==2.0.10
pyzmq==26.0.3
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.3
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
SpeechRecognition==3.9.0
sqlparse==0.4.4
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
tk==0.1.0
tornado==6.2
tqdm==4.66.4
traitlets==5.9.0
twilio==9.0.5
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
whitenoise==6.5.0
wikipedia==1.4.0
winshell==0.6
wolframalpha==5.0.0
wrapt==1.16.0
xmltodict==0.13.0
yarl==1.9.4
zipp==3.15.0
